Refactor GetAppServer and GetDatabaseServer api endpoints to remove MarshalOptions argument.
