+++ lib/services/resource.go
@@ -23,9 +23,50 @@ import (
 	"sync"
 	"time"
 
-	"github.com/gravitational/teleport/api/constants"
+	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/trace"
 )
+
+// MarshalConfig specifies marshalling options
+type MarshalConfig struct {
+	// Version specifies particular version we should marshal resources with
+	Version string
+
+	// SkipValidation is used to skip schema validation.
+	SkipValidation bool
+
+	// ID is a record ID to assign
+	ID int64
+
+	// PreserveResourceID preserves resource IDs in resource
+	// specs when marshaling
+	PreserveResourceID bool
+
+	// Expires is an optional expiry time
+	Expires time.Time
+}
+
+// GetVersion returns explicitly provided version or sets latest as default
+func (m *MarshalConfig) GetVersion() string {
+	if m.Version == "" {
+		return V2
+	}
+	return m.Version
+}
+
+// MarshalOption sets marshalling option
+type MarshalOption func(c *MarshalConfig) error
+
+// CollectOptions collects all options from functional arg and returns config
+func CollectOptions(opts []MarshalOption) (*MarshalConfig, error) {
+	var cfg MarshalConfig
+	for _, o := range opts {
+		if err := o(&cfg); err != nil {
+			return nil, trace.Wrap(err)
+		}
+	}
+	return &cfg, nil
+}
 
 // AddOptions adds marshal options and returns a new copy
 func AddOptions(opts []MarshalOption, add ...MarshalOption) []MarshalOption {
@@ -537,7 +578,7 @@ const V2SchemaTemplate = `{
 }`
 
 // MetadataSchema is a schema for resource metadata
-var MetadataSchema = fmt.Sprintf(baseMetadataSchema, constants.LabelPattern)
+var MetadataSchema = fmt.Sprintf(baseMetadataSchema, types.LabelPattern)
 
 // DefaultDefinitions the default list of JSON schema definitions which is none.
 const DefaultDefinitions = ``
